models package refactored
