+++ src/main/java/com/configcat/Result.java
@@ -1,0 +1,27 @@
+package com.configcat;
+
+public final class Result<T> {
+    private final T value;
+    private final String error;
+
+    private Result(T value, String error) {
+        this.value = value;
+        this.error = error;
+    }
+
+    T value() {
+        return this.value;
+    }
+
+    String error() {
+        return this.error;
+    }
+
+    static <T> Result<T> error(String error, T value) {
+        return new Result<>(value, error);
+    }
+
+    static <T> Result<T> success(T value) {
+        return new Result<>(value, null);
+    }
+}
