+++ src/main/java/com/configcat/models/Config.java
@@ -1,0 +1,27 @@
+package com.configcat.models;
+
+import com.google.gson.annotations.SerializedName;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class Config {
+    @SerializedName(value = "p")
+    private Preferences preferences;
+    @SerializedName(value = "f")
+    private Map<String, Setting> entries = new HashMap<>();
+
+    public Preferences getPreferences() {
+        return preferences;
+    }
+
+    public Map<String, Setting> getEntries() {
+        return entries;
+    }
+
+    boolean isEmpty() {
+        return EMPTY.equals(this);
+    }
+
+    public static final Config EMPTY = new Config();
+}
