+++ src/main/java/com/configcat/models/PercentageRule.java
@@ -1,4 +1,4 @@
-package com.configcat;
+package com.configcat.models;
 
 import com.google.gson.JsonElement;
 import com.google.gson.annotations.SerializedName;
