+++ src/test/java/com/configcat/ConfigFetcherTest.java
@@ -3,6 +3,8 @@ package com.configcat;
 import com.configcat.hooks.ConfigCatHooks;
 import com.configcat.log.ConfigCatLogger;
 import com.configcat.log.LogLevel;
+import com.configcat.models.Config;
+import com.configcat.models.Entry;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import okhttp3.OkHttpClient;
